+++ NEWS.md
@@ -12,6 +12,7 @@ New features
   attempts to generate an appropriate matching (piecepack stackpack) subpack and hexpack (#161).
 * ``pp_cfg`` objects ``get_suit_color`` function is now vectorized 
   (and by default now returns the suit colors for each suit).
+* New function ``file2grob`` that imports a given image file as a grob.
 
 Bug fixes and minor improvements
 --------------------------------
@@ -22,7 +23,7 @@ Bug fixes and minor improvements
   when width and height are not equal.
 * Fixes bug in ``op_transform`` for inferring width/height of matchsticks 
   and shape/depth of pyramids.
-* Fixes bug in printing 'pp_cfg' objects when a custom grob function had been set.
+* Fixes bug in printing ``pp_cfg`` objects when a custom grob function had been set.
 
 piecepackr 1.1.1
 ================